% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigned_process.R
\name{assigned_process}
\alias{assigned_process}
\title{Identifying ecological associations as environmental filtering and dispersal limitation}
\usage{
assigned_process(
  link_table_row,
  OTUabd,
  p = 0.05,
  data,
  cutoff = 0,
  method = c("dl", "ef")
)
}
\arguments{
\item{OTUabd}{}

\item{p}{}

\item{data}{}

\item{cutoff}{}

\item{method}{}
}
\description{
Identifying ecological associations as environmental filtering and dispersal limitation
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
