% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_alphacon.R
\name{cal_alphacon}
\alias{cal_alphacon}
\title{Calculating the contributions of microbial interaction on alpha diversity}
\usage{
cal_alphacon(env_data = data, method = c("forward.sel", "olslm"), ps = ps)
}
\arguments{
\item{method}{}

\item{ps}{}
}
\value{

}
\description{
Calculating the contributions of microbial interaction on alpha diversity
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
