% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_betacon.R
\name{cal_betacon}
\alias{cal_betacon}
\title{Calculating the contributions of microbial interaction on beta diversity}
\usage{
cal_betacon(env_data = data, otu_abd = otu, method = c("pearson", "spearman"))
}
\arguments{
\item{otu_abd}{}

\item{method}{}
}
\value{

}
\description{
Calculating the contributions of microbial interaction on beta diversity
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
