% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_network.R
\name{cal_network}
\alias{cal_network}
\title{Construct ecological networks using sequencing-based community data}
\usage{
cal_network(
  ps = filteredps,
  N = 0,
  r.threshold = 0.6,
  p.threshold = 0.05,
  lambda.min.ratio = 0.01,
  nlambda = 20,
  ncores = 1,
  method = c("spearman", "pearson", "spieceasi")
)
}
\arguments{
\item{N}{}

\item{r.threshold}{}

\item{p.threshold}{}

\item{lambda.min.ratio}{}

\item{nlambda}{}

\item{ncores}{}

\item{method}{}
}
\description{
Construct ecological networks using sequencing-based community data
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
