% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ps.R
\name{filter_ps}
\alias{filter_ps}
\title{Filter OTU table by occurrence and abundance}
\usage{
filter_ps(ps = ps, occurrence.threshold = 0, abundance.threshold = 0)
}
\arguments{
\item{occurrence.threshold}{}

\item{abundance.threshold}{}
}
\value{

}
\description{
Filter OTU table by occurrence and abundance
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
