% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idirect.R
\name{idirect}
\alias{idirect}
\title{Disentangling direct from indirect relationships in association networks}
\usage{
idirect(pathname= NULL, filename= NULL,  outputname= NULL)
}
\arguments{
\item{pathname}{}

\item{filename}{}

\item{outputname}{}
}
\description{
idirect addresses the three main problems of ill-conditioning, self-loop and interaction strength overflow that may exist in previous methods 
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}