% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.R
\name{neg.mean}
\alias{neg.mean}
\title{Create function that averages only negative values in a vector}
\usage{
neg.mean(vector)
}
\arguments{
\item{vector}{}
}
\description{
Create function that averages only negative values in a vector
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
