% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmt.r
\name{nnsd}
\alias{nnsd}
\title{Random Matrix Theory to filter correlation coefficient}
\usage{
nnsd(sp)
}
\arguments{
\item{lcor}{}

\item{hcor}{}
}
\description{
Random Matrix Theory to filter correlation coefficient
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
