% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{pos.mean}
\alias{pos.mean}
\title{Create function that averages only positive values in a vector}
\usage{
pos.mean(vector)
}
\arguments{
\item{vector}{}
}
\value{

}
\description{
Create function that averages only positive values in a vector
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
