% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcmi.R
\name{qcmi}
\alias{qcmi}
\title{Quantify the local intensity of microbial biotic interaction}
\usage{
qcmi(igraph, OTU, pers.cutoff = 0)
}
\arguments{
\item{OTU}{}

\item{pers.cutoff}{}
}
\value{

}
\description{
Quantify the local intensity of microbial biotic interaction
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
