% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmt.r
\name{rmt}
\alias{rmt}
\title{Random Matrix Theory to filter correlation coefficient}
\usage{
rmt(cormat, lcor = 0.6, hcor = 0.8)
}
\arguments{
\item{lcor}{}

\item{hcor}{}
}
\description{
Random Matrix Theory to filter correlation coefficient
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
