% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_link_env.R
\name{test_link_env}
\alias{test_link_env}
\title{Classifying ecological associations into the process of environmental filtering}
\usage{
test_link_env(link_table_row, OTUabd, envdist)
}
\arguments{
\item{link_table_row}{}

\item{OTUabd}{}

\item{envdist}{}
}
\description{
Classifying ecological associations into the process of environmental filtering
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
