% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_network.R
\name{cal_network}
\alias{cal_network}
\title{Correlation network calculation of microbial community data}
\usage{
cal_network(
  ps = filteredps,
  N = 0,
  r.threshold = 0.6,
  p.threshold = 0.05,
  method = c("spearman", "pearson")
)
}
\arguments{
\item{N}{}

\item{r.threshold}{}

\item{p.threshold}{}

\item{method}{}
}
\value{

}
\description{
Correlation network calculation of microbial community data
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
