% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnsd.r
\name{nnsd}
\alias{nnsd}
\title{Accessory function in Random Matrix Theory}
\usage{
nnsd(sp)
}
\arguments{
\item{sp}{}
}
\description{
Accessory function in Random Matrix Theory
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
