% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcmi.R
\name{qcmi}
\alias{qcmi}
\title{Quantify the strength of putative biotic associations at the community level}
\usage{
qcmi(igraph, OTU, pers.cutoff = 0)
}
\arguments{
\item{OTU}{}

\item{pers.cutoff}{}
}
\description{
Quantify the strength of putative biotic associations at the community level
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
