% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_link_dl.R
\name{test_link_dl}
\alias{test_link_dl}
\title{Classifying ecological associations of dispersal limitation}
\usage{
test_link_dl(link_table_row, OTUabd, geodist)
}
\arguments{
\item{link_table_row}{}

\item{OTUabd}{}

\item{geodist}{}
}
\value{

}
\description{
Classifying ecological associations of dispersal limitation
}
\references{

}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
