% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_ps.R
\name{trans_ps}
\alias{trans_ps}
\title{Convert the data  to phyloseq format, including OTU table and taxa table}
\usage{
trans_ps(
  otu_table = otu_table,
  taxa_table = taxa_table,
  sample_data = NULL,
  phylo = NULL
)
}
\description{
Convert the data  to phyloseq format, including OTU table and taxa table
}
\author{
Contact: Xu Liu \email{xliu@issas.ac.cn}
}
